add NativeCentralManagerCallbacks
## なぜこのクラスを作る必要があったのか
NativeCentralManagerProxyがAOT.MonoPInvokeCallbackを参照しないようにするため。

これを参照してしまうと、DocFXを使う際にUnityEngineを参照してしまう
