+++ Packages/com.teach310.core-bluetooth-for-unity/Runtime/Initializer.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+namespace CoreBluetooth
+{
+    internal class Initializer
+    {
+        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
+        static void Initialize()
+        {
+            ServiceLocator.Instance.Register<INativeCentralManagerCallbacks>(new NativeCentralManagerCallbacks());
+        }
+    }
+}
