+++ Packages/com.teach310.core-bluetooth-for-unity/Runtime/Initializer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac3bce9747e914f5980212a50bfb4e48
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
