+++ Packages/com.teach310.core-bluetooth-for-unity/Runtime/NativeCallbacks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4a8784e988b834584b3d2e5e2e6b074e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
