+++ Packages/com.teach310.core-bluetooth-for-unity/Runtime/NativeCallbacks/NativeCentralManagerCallbacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cebdd2a683bf44a3d8287b2d184259bb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
