+++ Packages/com.teach310.core-bluetooth-for-unity/Runtime/NativeCentralManagerProxy.cs
@@ -1,24 +1,27 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.InteropServices;
 
 namespace CoreBluetooth
 {
+    internal interface INativeCentralManagerCallbacks
+    {
+        void Register(SafeNativeCentralManagerHandle centralPtr, INativeCentralManagerDelegate centralManagerDelegate);
+        void Unregister(SafeNativeCentralManagerHandle centralPtr);
+    }
+
     internal class NativeCentralManagerProxy : IDisposable
     {
-        readonly static Dictionary<IntPtr, INativeCentralManagerDelegate> s_centralManagerDelegateMap = new Dictionary<IntPtr, INativeCentralManagerDelegate>();
-
         readonly SafeNativeCentralManagerHandle _handle;
+        readonly INativeCentralManagerCallbacks _callbacks;
 
         public NativeCentralManagerProxy(SafeNativeCentralManagerHandle handle, INativeCentralManagerDelegate centralManagerDelegate)
         {
             _handle = handle;
+            _callbacks = ServiceLocator.Instance.Resolve<INativeCentralManagerCallbacks>();
             if (centralManagerDelegate != null)
             {
-                s_centralManagerDelegateMap[handle.DangerousGetHandle()] = centralManagerDelegate;
+                _callbacks.Register(handle, centralManagerDelegate);
             }
-            RegisterHandlers();
         }
 
         public void Connect(CBPeripheral peripheral)
@@ -62,65 +65,9 @@ namespace CoreBluetooth
             return NativeMethods.cb4u_central_manager_is_scanning(_handle);
         }
 
-        void RegisterHandlers()
-        {
-            NativeMethods.cb4u_central_manager_register_handlers(
-                _handle,
-                DidConnect,
-                DidDisconnectPeripheral,
-                DidFailToConnect,
-                DidDiscoverPeripheral,
-                DidUpdateState
-            );
-        }
-
         public void Dispose()
         {
-            s_centralManagerDelegateMap.Remove(_handle.DangerousGetHandle());
-        }
-
-        static INativeCentralManagerDelegate GetDelegate(IntPtr centralPtr)
-        {
-            return s_centralManagerDelegateMap.GetValueOrDefault(centralPtr);
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UCentralManagerDidConnectHandler))]
-        internal static void DidConnect(IntPtr centralPtr, IntPtr peripheralIdPtr)
-        {
-            GetDelegate(centralPtr)?.DidConnect(Marshal.PtrToStringUTF8(peripheralIdPtr));
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UCentralManagerDidDisconnectPeripheralHandler))]
-        public static void DidDisconnectPeripheral(IntPtr centralPtr, IntPtr peripheralIdPtr, int errorCode)
-        {
-            GetDelegate(centralPtr)?.DidDisconnectPeripheral(
-                Marshal.PtrToStringUTF8(peripheralIdPtr),
-                CBError.CreateOrNullFromCode(errorCode)
-            );
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UCentralManagerDidFailToConnectHandler))]
-        public static void DidFailToConnect(IntPtr centralPtr, IntPtr peripheralIdPtr, int errorCode)
-        {
-            GetDelegate(centralPtr)?.DidFailToConnect(
-                Marshal.PtrToStringUTF8(peripheralIdPtr),
-                CBError.CreateOrNullFromCode(errorCode)
-            );
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UCentralManagerDidDiscoverPeripheralHandler))]
-        public static void DidDiscoverPeripheral(IntPtr centralPtr, IntPtr peripheralPtr, int rssi)
-        {
-            GetDelegate(centralPtr)?.DidDiscoverPeripheral(
-                new SafeNativePeripheralHandle(peripheralPtr),
-                rssi
-            );
-        }
-
-        [AOT.MonoPInvokeCallback(typeof(NativeMethods.CB4UCentralManagerDidUpdateStateHandler))]
-        public static void DidUpdateState(IntPtr centralPtr, CBManagerState state)
-        {
-            GetDelegate(centralPtr)?.DidUpdateState(state);
+            _callbacks.Unregister(_handle);
         }
     }
 }
